added Datepicker component to newly cloned repo
