--- client/src/components/Calendar.jsx
@@ -1,28 +1,0 @@
-import React, {useState} from "react"; 
-import Datepicker from "react-tailwindcss-datepicker"; 
-
-export default function Calendar() { 
-
-const [value, setValue] = useState({ 
-
-startDate: new Date(), 
-endDate: new Date().setMonth(11) 
-
-}); 
-
-const handleValueChange = (newValue) => {
-console.log("newValue:", newValue); 
-setValue(newValue); 
-
-} 
-
-return (
-<Datepicker 
-
-value={value} 
-onChange={handleValueChange} 
-
-/> 
-
-);
-};
